requirements.txt
torch>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.9.0
pyyaml>=6.0
tqdm>=4.66.0
